:truck: Refs #711 - Move stuf zds to be in stuf package
