+++ src/openforms/conf/base.py
@@ -144,6 +144,7 @@ INSTALLED_APPS = [
     "zgw_consumers",
     "stuf",
     "stuf.stuf_bg",
+    "stuf.stuf_zds",
     "mozilla_django_oidc",
     "mozilla_django_oidc_db",
     "django_filters",
@@ -167,7 +168,7 @@ INSTALLED_APPS = [
     "openforms.registrations.contrib.demo",
     "openforms.registrations.contrib.zgw_apis",
     "openforms.registrations.contrib.email",
-    "openforms.registrations.contrib.stuf_zds",
+    "openforms.registrations.contrib.stuf_zds.apps.StufZDSPlugin",
     "openforms.registrations.contrib.objects_api",
     "openforms.prefill",
     "openforms.prefill.contrib.demo.apps.DemoApp",
